\name{CollocInfer-internal}
\title{CollocInfer internal functions}
\description{Internal undocumentation functions}
\usage{blocks2mat(H)}
\alias{SplineCoefsDC2sparse}
\alias{ProfileSSE.AllPar}
\alias{ProfileErr.AllPar}
\alias{ProfileDP.AllPar}
\alias{blocks2mat}
\alias{trimr}
\alias{Newey.West}
\alias{NeweyWest.Var}
\alias{checkweights}
\alias{check.lik.proc.data.coefs}
\alias{oderhs}
\alias{chemo.ode}
\alias{ProfileErr1}
\alias{ProfileDP1}
% \alias{coefs2list}
% \alias{list2diag}
% \alias{whichpars.default}
% \alias{index2logical}
% \alias{splitfn}
% docclass is function
\keyword{internal}
