\name{CollocInfer-package}
\alias{CollocInfer-package}
\alias{CollocInfer}
\docType{package}
\title{Collocation Inference in R}
\description{
  Functions carry out collocation inference method for nonlinear continuous-time
  dynamic systems. These are based on basis-expansion representations for the
  state of the system. Gradient-matching, profiling and EM algorithms are
  supported.
}
\details{
  \tabular{ll}{
    Package: \tab CollocInfer\cr
    Type: \tab Package\cr
    Version: \tab 2.1.0\cr
    Date: \tab 2009-08-19\cr
    License: \tab GPL-2\cr
    LazyLoad: \tab yes\cr
  }
}
\author{
  Giles Hooker,  Luo Xiao

  Maintainer:  Giles Hooker <giles.hooker@cornell.edu>
}
\references{
  Ramsay, James O., Giles Hooker, Jiguo Cao and David Campbell (2007), "Parameter
  Estimation in Ordinary Differential Equations: A Generalized Smoothing
  Approach", \emph{Journal of the Royal Statistical Society}, 69
  
  Ramsay, James O., and Silverman, Bernard W. (2006), \emph{Functional
    Data Analysis, 2nd ed.}, Springer, New York.
}

