\name{FhNest}
\alias{FhNestPars}
\alias{FhNestCoefs}
\title{Estimated Parameters for FitzHugh-Nagumo data}
\description{Parameters Estimated for FhN Data -- used to speed up examples}
\usage{
FhNest
}
\format{
\itemize{
\item{FhNestPars}{ Estimated parameters for the FhN Data example.}
\item{FhNestCoefs}{ Estimated coefficients for the FhN Data example.}
}}
\source{
James Ramsay, Giles Hooker David Campbell and Jiguo Cao, 2007.
"Parameter Estimation for Differential Equations: A Generalized Smoothing Approach".
Journal of the Royal Statistical Society Vol 69 No 5.
}
