\name{NSdata}
\alias{NSgroundwater}
\alias{NSrainfall}
\alias{NStimes}
\title{North Shore data}
\description{Groundwater Data from Vancouver's North Shore}
\usage{
NSgroundwater
}
\format{
\itemize{
\item{NSgroundwater}{ A 315 by 1 matrix of data on groundwater level collected in vancouver.}
\item{NStimes}{ A vector of 315 observation times corresponding to NSgroundwater.}
\item{NSrainfall}{ Rainfall as a covariate to NSgroundwater; this quantity is lagged by 3 days. }
}}
